+++ src/enhancements.js
@@ -1,3 +1,38 @@
+const srcset = require('srcset');
+const URL = require('url').URL;
+const replaceElementType = require('./replace-element-type');
+
+/* 
+	Convert AMP markup to HMTL markup
+	(naive / incomplete implementation)
+*/
+function ampToHtml(doc) {
+	// Convert <amp-img> to <img>
+	Array.from(doc.querySelectorAll('amp-img')).forEach(ampImg => {
+		replaceElementType(ampImg, 'img', doc);
+	});
+}
+
+function fixLazyLoadedImages(doc) {
+	Array.from(
+		doc.querySelectorAll(`
+		img[data-src], 
+		img[data-srcset],
+		img[data-sizes]
+	`)
+	).forEach(img => {
+		['src', 'srcset', 'sizes'].forEach(attr => {
+			if (attr in img.dataset) {
+				img.setAttribute(attr, img.dataset[attr]);
+			}
+		});
+	});
+
+	Array.from(doc.querySelectorAll('[loading="lazy"]')).forEach(el => {
+		el.removeAttribute('loading');
+	});
+}
+
 function imagesAtFullSize(doc) {
 	/*
 		Replace:
@@ -8,12 +43,30 @@ function imagesAtFullSize(doc) {
 		With:
 			<img src='original-size.png'/>
 	 */
+	let include_pattern = /\.(png|jpg|jpeg|gif|svg)$/i;
+	let exclude_patterns = [
+		/*
+			Exclude Wikipedia links to image file pages
+		*/
+		/wiki\/File\:/,
+
+		/* 
+			Exclude images embedded in Markdown files
+			hosted on github.com.
+			See: https://github.com/danburzo/percollate/issues/84
+		*/
+		/github\.com/
+	];
+
 	Array.from(doc.querySelectorAll('a > img:only-child')).forEach(img => {
 		let anchor = img.parentNode;
 		let original = anchor.href;
 
 		// only replace if the HREF matches an image file
-		if (original.match(/\.(png|jpg|jpeg|gif|svg)$/)) {
+		if (
+			include_pattern.test(original) &&
+			!exclude_patterns.some(pattern => pattern.test(original))
+		) {
 			img.setAttribute('src', original);
 			anchor.parentNode.replaceChild(img, anchor);
 		}
@@ -41,25 +94,101 @@ function wikipediaSpecific(doc) {
 	).forEach(el => el.remove());
 }
 
+/* 
+	Mark some links as not needing their HREF appended.
+*/
 function noUselessHref(doc) {
-	/* 
-		Mark some links as not needing their HREF appended:
-		- links whose text content is the HREF
-		- in-page anchors
-	*/
+	Array.from(doc.querySelectorAll('a'))
+		.filter(function (el) {
+			let href = el.getAttribute('href') || '';
 
-	Array.from(doc.querySelectorAll(`a`))
-		.filter(function(el) {
-			return (
-				(el.getAttribute('href') || '').match(/^\#/) ||
-				el.getAttribute('href') === el.textContent.trim()
-			);
+			// in-page anchors
+			if (href.match(/^\#/)) {
+				return true;
+			}
+
+			let textContent = el.textContent.trim();
+
+			// links whose text content is the HREF
+			// or which don't have any content.
+			return !textContent || textContent === href;
 		})
 		.forEach(el => el.classList.add('no-href'));
 }
 
+/*
+	Convert relative URIs to absolute URIs:
+	
+	* the `href` attribute of <a> elements (except for in-page anchors)
+	* the `src` attribute of <img> elements
+	* the `srcset` attribute of <source> elements (inside <picture> elements)
+ */
+function relativeToAbsoluteURIs(doc) {
+	function absoluteSrcset(str) {
+		return srcset.stringify(
+			srcset.parse(str).map(item => ({
+				...item,
+				url: new URL(item.url, doc.baseURI).href
+			}))
+		);
+	}
+
+	Array.from(doc.querySelectorAll('a:not([href^="#"])')).forEach(el => {
+		el.setAttribute('href', el.href);
+	});
+
+	Array.from(doc.querySelectorAll('img')).forEach(el => {
+		el.setAttribute('src', el.src);
+	});
+
+	Array.from(
+		doc.querySelectorAll('picture source[srcset], img[srcset]')
+	).forEach(el => {
+		el.setAttribute('srcset', absoluteSrcset(el.getAttribute('srcset')));
+	});
+}
+
+/*
+	Wraps single images into <figure> elements,
+	adding the image's `alt` attribute as <figcaption>
+ */
+function singleImgToFigure(doc) {
+	Array.from(doc.querySelectorAll('img:only-child')).forEach(image => {
+		let fig = doc.createElement('figure');
+		fig.appendChild(image.cloneNode());
+
+		let alt = image.getAttribute('alt');
+		if (alt) {
+			let figcaption = doc.createElement('figcaption');
+			figcaption.textContent = alt;
+			fig.appendChild(figcaption);
+		}
+
+		// Replace paragraph with figure
+		if (image.parentNode.matches('p') && image.parentNode.parentNode) {
+			image.parentNode.parentNode.replaceChild(fig, image.parentNode);
+		} else {
+			image.parentNode.replaceChild(fig, image);
+		}
+	});
+}
+
+/*
+	Expands <details> elements
+*/
+function expandDetailsElements(doc) {
+	Array.from(doc.querySelectorAll('details')).forEach(el =>
+		el.setAttribute('open', true)
+	);
+}
+
 module.exports = {
+	ampToHtml,
+	fixLazyLoadedImages,
 	imagesAtFullSize,
 	noUselessHref,
-	wikipediaSpecific
+	wikipediaSpecific,
+	relativeToAbsoluteURIs,
+	singleImgToFigure,
+	expandDetailsElements
 };